Close Phase 9D: PreTemporada/Temporada/Encerramento lifecycle, v33/v34 migrations, integration test suite
- Unified calendar model: 74 entries, season_week axis (sw 10-51), season_phase=Temporada
- v33 migration: adds season_week column, remaps ThematicSlot, Finalizada→Encerramento
- v34 migration: converts BlocoRegular→Temporada, generates partial production/endurance entries
- skip_all_pending_races_in_base_dir handles Temporada phase via simulate_category_race
- Integration tests (tests_9d.rs): 11 tests covering invariants, 2-season cycle, legacy traversal
- Player endurance LMP2 path now covered via contract surgery test
- Removed export/, market/models, season command, legacy race/simulation modules
- DESIGN.md: Phase 9D documented, schema v34, season_week maintenance rule added

+++ src-tauri/src/evolution/rookies.rs
@@ -71,7 +71,7 @@ fn generate_single_rookie(
         identity.nome_completo,
         identity.nacionalidade_label,
         identity.genero,
-        age as u32,
+        age,
         ano_inicio,
     );
     driver.categoria_atual = None;
